Bugfix for TaggedButtonGrid in combination with AutoCleanForm

+++ TelegramBotBase/Controls/Hybrid/TaggedButtonGrid.cs
@@ -455,7 +455,7 @@ public class TaggedButtonGrid : MultiView
                     if (result.MessageText == "📁")
                     {
                         //Remove button click message
-                        if (DeletePreviousMessage)
+                        if (DeletePreviousMessage && !Device.ActiveForm.IsAutoCleanForm())
                         {
                             await Device.DeleteMessage(result.MessageId);
                         }
